chats_front/requirements.txt
langchain
langserve
python-dotenv
pydantic
openai
sse_starlette
fastapi
uvicorn